Heavy valgrind session: removed memory leak, replaced some stl containers with arrays (configurable in bsConfig.h), changed script stack from uint32s to floats (much faster).

+++ include/bsScriptMachine.h
@@ -50,7 +50,7 @@ namespace BS_NMSP
 	typedef void (*ErrorFunction)(const char*);
 
 	class BulletGunBase;
-	typedef int (*FireFunction)(BulletGunBase*, float, float, uint32*);
+	typedef int (*FireFunction)(BulletGunBase*, float, float, float*);
 
 	class BulletMachineBase;
